services/ai-hedge-fund/pyproject.toml
[tool.poetry]
name = "gicm-ai-hedge-fund"
version = "1.0.0"
description = "AI Hedge Fund with famous investor personas - Crypto/Solana edition"
authors = ["gICM"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# LangChain ecosystem
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langgraph = "^0.2.0"

# Core dependencies
pydantic = "^2.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"

# API & CLI
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
typer = {extras = ["all"], version = "^0.12.0"}
rich = "^13.7.0"

# Data processing
pandas = "^2.2.0"
numpy = "^1.26.0"

# Database
supabase = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
black = "^24.0.0"
ruff = "^0.4.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
gicm-hedge = "src.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
